+++ e3nn_jax/_src/fc_tp_haiku.py
@@ -1,7 +1,8 @@
 import haiku as hk
+import jax.numpy as jnp
 
 import e3nn_jax as e3nn
-from e3nn_jax._src.tensor_products import naive_broadcast_decorator
+from e3nn_jax.legacy import FunctionalFullyConnectedTensorProduct
 
 
 class FullyConnectedTensorProduct(hk.Module):
@@ -15,6 +16,13 @@ class FullyConnectedTensorProduct(hk.Module):
     def __call__(
         self, x1: e3nn.IrrepsArray, x2: e3nn.IrrepsArray, **kwargs
     ) -> e3nn.IrrepsArray:
+        x1 = e3nn.as_irreps_array(x1)
+        x2 = e3nn.as_irreps_array(x2)
+
+        leading_shape = jnp.broadcast_shapes(x1.shape[:-1], x2.shape[:-1])
+        x1 = x1.broadcast_to(leading_shape + (-1,))
+        x2 = x2.broadcast_to(leading_shape + (-1,))
+
         if self.irreps_in1 is not None:
             x1 = x1.rechunk(self.irreps_in1)
         if self.irreps_in2 is not None:
@@ -23,7 +31,7 @@ class FullyConnectedTensorProduct(hk.Module):
         x1 = x1.remove_zero_chunks().simplify()
         x2 = x2.remove_zero_chunks().simplify()
 
-        tp = e3nn.FunctionalFullyConnectedTensorProduct(
+        tp = FunctionalFullyConnectedTensorProduct(
             x1.irreps, x2.irreps, self.irreps_out.simplify()
         )
         ws = [
@@ -37,8 +45,10 @@ class FullyConnectedTensorProduct(hk.Module):
             )
             for ins in tp.instructions
         ]
-        f = naive_broadcast_decorator(
-            lambda x1, x2: tp.left_right(ws, x1, x2, **kwargs)
-        )
+        f = lambda x1, x2: tp.left_right(ws, x1, x2, **kwargs)
+
+        for _ in range(len(leading_shape)):
+            f = e3nn.utils.vmap(f)
+
         output = f(x1, x2)
         return output.rechunk(self.irreps_out)