+++ e3nn_jax/legacy.py
@@ -1,0 +1,13 @@
+from e3nn_jax._src.legacy.core_tensor_product import (
+    FunctionalTensorProduct,
+    Instruction,
+)
+from e3nn_jax._src.legacy.functional_fully_connected_tensor_product import (
+    FunctionalFullyConnectedTensorProduct,
+)
+
+__all__ = [
+    "Instruction",
+    "FunctionalTensorProduct",
+    "FunctionalFullyConnectedTensorProduct",
+]